cover case where a user creates a resource of their own, introducing a 'principal user' (naive implementation)
